fix: migrate verify
